requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pyyaml==6.0.2
mysql-connector-python==9.0.0
python-dotenv==1.0.1
aiofiles==24.1.0
pydantic==2.9.2
pydantic-settings==2.5.2
